dodano autentykację i autoryzację

+++ RestaurantApi/Services/RestaurantService.cs
@@ -1,13 +1,17 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 
 using AutoMapper;
 
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 
+using RestaurantApi.Autorization;
 using RestaurantApi.Entities;
 using RestaurantApi.Exceptions;
+using RestaurantApi.Middleware;
 using RestaurantApi.Models;
 
 namespace RestaurantApi.Services
@@ -28,12 +32,17 @@ namespace RestaurantApi.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
 
         public void Modify(int id, ModifyRestaurantDto modDto)
@@ -42,7 +51,14 @@ namespace RestaurantApi.Services
                    .Restaurants
                    .FirstOrDefault(r => r.Id == id);
 
-            if (restaurant == null) throw new NotFoundException("Restaurant not found");
+            if (restaurant == null)
+                throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
 
             restaurant.Name = modDto.Name;
             restaurant.Description = modDto.Description;
@@ -59,6 +75,12 @@ namespace RestaurantApi.Services
                    .Restaurants
                    .FirstOrDefault(r => r.Id == id);
             if (restaurant == null) throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                     new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResult.Succeeded)
+                throw new ForbidException();
 
             _dbContext.Restaurants.Remove(restaurant);
             _dbContext.SaveChanges();
@@ -93,6 +115,7 @@ namespace RestaurantApi.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
             return restaurant.Id;
